% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_flowjo_2020.R
\name{parse_flowjo_2020}
\alias{parse_flowjo_2020}
\title{Parse_flowjo_2020  function}
\usage{
parse_flowjo_2020(files_path, wsp_file, group = "All Samples", plot = FALSE)
}
\arguments{
\item{files_path}{a path  to the directory containing the FCS files}

\item{wsp_file}{a path to the workspace}

\item{group}{Defaults to "All Samples"}

\item{plot}{Defaults to FALSE}
}
\description{
This function replaces parse_flowjo() from FlowSOM package.
}
\examples{
files_path <- "/Users/christophe/R/2019R/BM84-2/ CA/ CA 2020/FSOM 18"
wsp_path <- "/Users/christophe/R/2019R/BM84-2/ CA/ CA 2020/FSOM 18/BM84-2 FS18.wsp"
tubes  <- parse_flowjo_2020(files_path, wsp_path, plot = TRUE)
parse_flowjo_2020()
}
